New noteModule module
This model contains high-level interfaces to add and update notes.
It defines two classes, Note and NoteSet. NoteSet is intended to be a collection of Note(s).
Currently every Note has various methods to add it and update it on the anki deck.
NoteSet is intended to have methods to create Note(s) objects from a file by parsing it etc, and to be able to upload them all at the same time.