New logic to check connection
The check_connection function, and therefore the ensure_connectivity wrapper/decorator,
now write to an environment variable the last time where the actual check was made.
Subsequent calls to the function then check if the last check happened in the last ten
seconds, in which case they do not make the check again but they simply return true.
This is done to avoid checking multiple times when e.g. a collection of cards is being
uploaded at the same time.